projects/requirements.txt
google-adk>=1.18.0
litellm>=1.76.0
anthropic>=0.57.0
streamlit>=1.49.0
yfinance>=0.2.65
pandas>=2.3.0
numpy>=2.3.0
requests>=2.32.0
beautifulsoup4>=4.13.0
lxml>=6.0.0
feedparser>=6.0.11
sec-edgar-downloader>=5.0.3
python-dotenv>=1.1.0
vaderSentiment>=3.3.2
tenacity>=9.1.0
